update: refactor order and order product

+++ migrations/sqls/20240329025301-order-table-up.sql
@@ -1,5 +1,5 @@
 create table orders(
     id serial primary key,
-    user_id bigint references users(id),
+    user_id int references users(id),
     status varchar(10)
 );
